refactor: extract BackendVersion and BackendRow from manage-backends-modal
